refs #12192 Switched semesterNumber field type from Byte to Integer

+++ src/main/java/ru/nsu/ostest/adapter/in/rest/model/laboratory/LaboratoryCreationRequestDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public record LaboratoryCreationRequestDto(String name,
                                            String description,
-                                           Byte semesterNumber,
+                                           Integer semesterNumber,
                                            LocalDateTime deadline,
                                            Boolean isHidden) {
 }
