+++ src/main/java/ru/nsu/ostest/adapter/in/rest/model/laboratory/LaboratoryDto.java
@@ -9,7 +9,7 @@ public record LaboratoryDto(
         Long id,
         String name,
         String description,
-        Byte semesterNumber,
+        Integer semesterNumber,
         LocalDateTime deadline,
         Boolean isHidden,
         List<TestDto> tests
